requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
torch==2.1.0
transformers==4.35.0
faiss-cpu==1.7.4
pillow==10.1.0
python-multipart==0.0.6
numpy==1.24.3
scikit-learn==1.3.2

# AWS
boto3==1.28.44
sagemaker==2.193.0

# Development
pytest==7.4.3
black==23.10.1
flake8==6.1.0
pre-commit==3.5.0
pytest-cov==4.1.0